Remove declaration of non-existent method in iOS module
This "multiply" method is not implemented in
ThetaClientReactNative.swift. It was probably copied from a guide and
left by mistake.
Having this declaration with no corresponding definition causes a
crash at app start with React Native 0.76 and new architecture:
when looking for method signature for selector
"multiply:withB:withResolver:withRejecter:" in parseExportedMethods
from RCTInteropTurboModule.mm, we get nil which causes a crash right
after.